Move Markdown routes
