+++ aesops/__init__.py
@@ -16,6 +16,8 @@ from aesops import routes
 from data_models import model_store
 
 from .blueprints.login_blueprint import login_blueprint
+from .blueprints.markdown_blueprint import markdown_blueprint
 from .blueprints.tournament_blueprint import tournament_blueprint
 app.register_blueprint(login_blueprint)
+app.register_blueprint(markdown_blueprint)
 app.register_blueprint(tournament_blueprint)
