+++ aesops/templates/base.html
@@ -24,10 +24,10 @@
             <a class="nav-link" aria-current="page" href="{{ url_for('index') }}">Tournament Directory</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="{{ url_for('howto') }}">How To</a>
+            <a class="nav-link" href="{{ url_for('markdown.howto') }}">How To</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="{{ url_for('about') }}">About</a>
+            <a class="nav-link" href="{{ url_for('markdown.about') }}">About</a>
           </li>
         </ul>
         <ul class="navbar-nav mb-2 mb-lg-0">
